Tarefa4_2022li1g011_Gloss

Description

Módulo para a realização da Tarefa 4 do projeto de LI1 em 2022/23.

Função principal

jogoTerminou :: Jogo -> Bool

A função jogoTerminou que recebe um Jogo e uma jogada que entrega um Booleano sendo se for True o jogo terá terminado e se for False o jogo continua.

A função aplica animação ao Jogo através de 2 funções auxiliares principais, uma que movimenta o mapa e outra que faz moviementar o personagem, na função principal vai juntar as funções auxiliares de modo a criar um tipo de animação ao Jogo.

O moviemnto do jogador não acontece se e só se o local para onde ele se for mover for ocupado por uma árvore.

A função pode ser definida da seguinte forma:

jogoTerminou :: Jogo -> Bool
jogoTerminou jogo@(Jogo (Jogador (x,y)) (Mapa l ((t,(o:os)):r))) = jogadorForaMapa jogo || jogadorMortoObstaculo (Jogo (Jogador (x,y)) (Mapa l ((linhaY jogo):r)))

Exemplos de utilização

Exemplo 1 - Movimenta Para Cima

>>> animaJogo (Jogo (Jogador (1,2)) (Mapa 4 ([(Rio 2,[Tronco,Nenhum,Tronco,Nenhum]),(Estrada (-2),[Carro,Carro,Nenhum,Nenhum]),(Relva,[Arvore,Nenhum,Arvore,Nenhum]),(Rio 3,[Tronco,Nenhum,Tronco,Tronco])]))) (Move Cima)
Jogo (Jogador (1,1)) (Mapa 4 [(Rio 2,[Tronco,Nenhum,Tronco,Nenhum]),(Estrada (-2),[Nenhum,Nenhum,Carro,Carro]),(Relva,[Arvore,Nenhum,Arvore,Nenhum]),(Rio 3,[Nenhum,Tronco,Tronco,Tronco])])

Exemplo 2 - Movimenta Para Baixo (Com Uma Árvore em Baixo)

>>> animaJogo (Jogo (Jogador (2,1)) (Mapa 4 ([(Rio (-3),[Tronco,Nenhum,Tronco,Nenhum]),(Estrada (-2),[Carro,Carro,Nenhum,Nenhum]),(Relva,[Arvore,Nenhum,Arvore,Nenhum]),(Rio 1,[Tronco,Nenhum,Tronco,Tronco])]))) (Move Baixo)
Jogo (Jogador (2,1)) (Mapa 4 [(Rio (-3),[Nenhum,Tronco,Nenhum,Tronco]),(Estrada (-2),[Nenhum,Nenhum,Carro,Carro]),(Relva,[Arvore,Nenhum,Arvore,Nenhum]),(Rio 1,[Tronco,Tronco,Nenhum,Tronco])])

jogadorForaMapa :: Jogo -> Bool

jogadorMortoObstaculo :: Jogo -> Bool

linhaY :: Jogo -> (Terreno, [Obstaculo])

listaTerrenos1 :: [(Terreno, [Obstaculo])] -> [Terreno]